Merged PR 7772625: [gitops] Gitops Service
# Description

The `/gitops` directory contains code for the proposed gitops service for azure.

## Concepts

- RepoResource - The azure representation of a Github Repository
- CloudGitOpsResource - A gitops resource representing a ARM deployment
- EdgeGitOpsResource - A gitops resource representing an Edge deployment

## Helm
The `gitops/helm` contains the helm chart for deploying the GitOps  service to a kubernetes cluster.

## Operation
This service serves some API routes for starting and stopping GitOps operations.

When a cloud GitOps  resource is started, The service would fetch the template from the specified git repository that the operation belongs to and checks if that version of the file has been deployed to ARM. If it hasn't, the deployment is created or updated in ARM

When an Edge GitOps resource is started, The service would periodically fetch the template and parameters from the git repository, and creates an Edge deployment (currently a Symphony Instance) and Edge Templates (currently a Symphony Solutions) and deploys it to the specified targets

At the moment file sync is very rudimentary. Each GitOps resource runs at the specified interval and pulls the file content of interest from the git repository. If any of the content has changed since the last deploy, an update ARM deployment is created and applied.