Merged PR 8214549: Add Managed Identity suppport
## Overview
This adds support for system assigned managed identities to be used by the symphony API, no other change is needed to support

## TLDR
No extra work needs to be done on the symphony codebase to access azure resources. The Azure identity library would work out of the box for Managed Identities. (This is still susceptible to the chicken and egg problem where system assigned manage  identities usually don't have permissions assigned to it at creation time. They are usually assigned after the fact)

## Details
For resources like VMs and AKS clusters that natively support managed identities, azure has a web service that runs on the host to provide access to MSAL and IDAL libraries. The webserver is the Managed Identity representative on the host machine and these libraries by default call out to the webserver to retrieve a token from AAD..

To simulate this behavior on ARC connected kube clusters, The Extension Identity operator in the  Running in the cluster operates on two CRDS
- AzureExtensionIdentity
- AzureClusterIdentityRequest
With these two mechanisms arch extensions can retrieve AAD tokens. In the Azure Extension identity CRD you specify some gubernates service accounts and also a namespace. The name space is used for storing the token in the kubernetes secret store and the. The operator would also create a role and rolebinding that allows the specified service accounts access the created secret.

`Values.global.azure.identity` Has the spec for an msi-adapter container. We mount it as a sidecar so it's available as a local server to symphony-api. Whenever the MSAL libraries make a request to the msi-sidecar for a token, It creates (or reuses) an AzureClusterIdentityRequest. The ARC extension identity agent picks it up, fetches a token from AAD, creates/updates a secret in the secret store, updates the status of the `AzureClusterIdentityRequest` request with a reference to the secret. The Msal agent, having access to this object now knows where the secret is and how to retrieve it, It gets the access token and returns to MSAL.

Related work items: #24090916