Merged PR 8230451: Fix race condition between solution and instance
# Overview

Currently, an instance refers to a solution and if the solution doesn't exist on the cluster at the time of reconciliation, a reconciliation request is requeued 3 mins later in hopes that the solution might be there. This isn't ideal especially if the instance and solution are applied at roughly the same time but the instance reconciler is triggered first.

To combat this, the instance controller also watches the solution objects to to queue requests for any matching instances. The problem was it wasn't selecting instances correctly so it was never able to queue requests for matching instances.

This PR fixes the issue indexing the `spec.solution` field on the instance, and updating the typo when querying instances based on this field.

This also adds a change predicate so that the Reconcile is called only on object changes (and intentional Reconcile requests). By default, a reconcile request is queued whenever there is a change on a watched object, including changes in a sub resource. This means in the reconcile request, if there's a change in the status, the request is reconciled again. This isn't ideal especially because the Reconciler updates the status property during reconciliation and the `lastUpdated` field is almost guaranteed to be different on every reconciliation. The newly added change predicates would allow reconcile requests only when the core object has changed or annotations have changed.

Related work items: #24172256, #24172259