Merged PR 8226867: Fix Helm Provider config
# Overview

This change fixes initializing the helm provider. In a ideal world, we should be able to unmarshal the config map into the desired struct, but the config is currently defined as `map[string]string`.

This reverts the config initialization to re-support this.

Once support for `map[string]interface{}` is implemented, we can switch to marshalling.

Related work items: #24182846