Merged PR 8275695: Add finalizer removers and add delete job
Some of our objects (namely instances and targets) have finalizers on them that cause Kubernetes to add a `deletionTimestamp` on the object when a delete is triggered. This gives controllers the option to delete any external resources before the object is actually deleted. The object is deleted only when the finalizers are removed.

When we uninstall the Symphony extension, it deletes all the resources created by the helm release. Does not include the custom resources (CROs) created by the user. The CRDs are part of the release, but when a CRD is deleted it tries to delete all custom resources of that type.

But if the Operator is deleted before the CRD's get a chance to be deleted, the CRO's would be marked with the deletionTimestamp and never get deleted. So the uninstall would remain stuck

This PR adds a hook to trigger resource cleanup during uninstalls

Related work items: #17153896, #22132237, #23883319, #24260820, #24263389