Merged PR 8311243: [zipkin] Adding zipkin to Symphony as part of the arc extension
Changed behavior:
Before: copied the symphony-api.json over to docker image
After: mount volume, have config map to store the symphony-api.json config file so that we have the .Release name (extension name) and namespace during installation time, this way we know what url (dynamic url) it should be for zipkin to use/post

If .Values zipkin is needed, we simply deploy zipkin as dependency for symphony so that we have somewhere for traces to go to.

Tested this by installing symphony as helm chart
Tested this by pushing both symphony api and arc extension image to acr and install as arc extension, things are working. zipkin is deployed and symphony is posting traces to zipkin url which is also working.

Related work items: #24310838