Merged PR 8437007: Add component level status
This PR is the initial commit to enable component level status reporting.

We need to make change for each provider to have a correct reflection of component status, similar to what I did in this PR for the Helm provider. All the other providers now always return `Untouched` for each component.
![image (2).png](https://msazure.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/c0592da2-4ee2-4089-ac51-f9ad1fedda3c/pullRequests/8437007/attachments/image%20%282%29.png)
![image.png](https://msazure.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/c0592da2-4ee2-4089-ac51-f9ad1fedda3c/pullRequests/8437007/attachments/image.png)
![image (3).png](https://msazure.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/c0592da2-4ee2-4089-ac51-f9ad1fedda3c/pullRequests/8437007/attachments/image%20%283%29.png)

Related work items: #24291772, #24510038