Include JSON config in Rust Target Provider constructor (#708)
Improved several formal aspects of the Rust binding.
In particular
* added doc comments,
* made the code comply with the rustfmt rules,
* replaced println statements with tracing events.

Also removed ITargetProvider::init in favor of providing the (full)
JSON configuration to the create_provider function that is used to
create the Target Provider instance. This accounts for the fact that
it is usually a little tedious to alter internal state of an instance in
Rust.